feat: Add Anno 117 (Pax Romana) support
- Add GameVersion.Anno117 enum value and preset tree filter
- Add AnnoObject properties: Rotation, RotationCenter, TileQuadrants, RoadInfluenceFactor
- Add BuildingInfo: RoadInfluenceFactor, BuildBlockerVariants, BuildBlocker changed to double
- Add SessionLayout/IslandLayout models, LayoutFile.Sessions for savegame import
- Add diagonal building rendering (RotateTransform in CanvasRenderer)
- Add weighted road influence factor in RoadSearchHelper BFS
- Add AnnoDesigner.Import project with full savegame reader (from Just-Chaldea fork)
- Add AnnoDesigner.Gamedata project for geometry/asset support
- Add ImportSavegameWindow (File > Import > Anno 117 Savegame)
- Add 264 Anno 117 building icons (A8_*.png)
- Update presets.json with 454 Anno 117 buildings
- Add MathHelper.GetDiagonalSize for diagonal grid sizing
- Add BuildingInfoExtensions.ToAnnoObject() parameterless overload

Based on Just-Chaldea/anno-designer fork (https://github.com/Just-Chaldea/anno-designer)
Refs: AnnoDesigner/anno-designer#459

+++ PresetParser/BuildingBlockProvider.cs
@@ -76,7 +76,7 @@ namespace PresetParser
                     return false;
                 }
 
-                building.BuildBlocker = new SerializableDictionary<int>();
+                building.BuildBlocker = new SerializableDictionary<double>();
 
                 //Convert the strings to a Variable and replace the "." for a "," to keep calculatable numbers
                 var xfNormal1 = double.Parse(node1["xf"]?.InnerText ?? "", NumberStyles.Any, CultureInfo.InvariantCulture);
@@ -191,7 +191,7 @@ namespace PresetParser
                 if (ifoDocument.FirstChild != null)
                 {
                     var node = ifoDocument.FirstChild[BUILDBLOCKER]?.FirstChild;
-                    building.BuildBlocker = new SerializableDictionary<int>();
+                    building.BuildBlocker = new SerializableDictionary<double>();
 
                     var x = 0;
                     var z = 0;
